+++ src/core/chart-components/legend-cat/buttons.js
@@ -1,23 +1,5 @@
 import extend from 'extend';
-
-function directionTriangle(x, y, r, d = 'bottom') {
-  if (d === 'left' || d === 'right') {
-    let right = d === 'right';
-    x += (right ? -r : r);
-    return `
-      M ${x} ${y - r}
-      L ${x} ${y + r}
-      L ${right ? x + (r * 1.5) : x - (r * 1.5)} ${y} Z
-    `;
-  }
-  let bottom = d === 'down';
-  y += (bottom ? -r : r);
-  return `
-    M ${x - r} ${y}
-    L ${x + r} ${y}
-    L ${x} ${bottom ? y + (r * 1.5) : y - (r * 1.5)} Z
-  `;
-}
+import triangle from '../../symbols/triangle';
 
 export default function createButton({ x, y, width, height, direction, action, rect, symbol }) {
   const buttonRect = extend({
@@ -32,14 +14,20 @@ export default function createButton({ x, y, width, height, direction, action, r
     opacity: 1
   }, rect);
 
-  const r = Math.min(width / 2, height / 2) * 0.6;
+  const r = Math.min(width, height) * Math.max(0, Math.min(symbol.size || 0.5, 1));
   const symX = x + (width / 2);
   const symY = y + (height / 2);
 
-  const buttonSymbol = extend({}, symbol, {
-    type: 'path',
-    d: directionTriangle(symX, symY, r, direction)
+  const buttonSymbol = triangle({
+    x: symX,
+    y: symY,
+    size: r,
+    direction
   });
+
+  buttonSymbol.fill = symbol.fill;
+  buttonSymbol.stroke = symbol.stroke;
+  buttonSymbol.strokeWidth = symbol.strokeWidth;
 
   const container = {
     type: 'container',